+++ src/test/java/com/mtvs/devlinkbackend/crud/GuildCRUDTest.java
@@ -1,0 +1,10 @@
+package com.mtvs.devlinkbackend.crud;
+
+import jakarta.transaction.Transactional;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+@Transactional
+public class GuildCRUDTest {
+
+}
